+++ files/private-chef-cookbooks/private-chef/recipes/default.rb
@@ -152,17 +152,27 @@ file "/etc/opscode/worker-private.pem" do
   content worker_key.to_pem.to_s unless File.exists?('/etc/opscode/worker-public.pem')
 end
 
+# If we are doing initial key generation,
+# generate a new key.
 unless File.exists?('/etc/opscode/pivotal.pem')
-  cert, key = OmnibusHelper.gen_certificate
+  key = OpenSSL::PKey::RSA.generate(2048)
 end
 
-file "/etc/opscode/pivotal.cert" do
-  owner "root"
-  group "root"
-  mode "0644"
-  content cert.to_s unless File.exists?('/etc/opscode/pivotal.pem')
+# This will be cleaned up during bootstrap.
+# Only generate if pivotal.pem doesn't exist,
+# because that means we are doing initial key generation
+# per the step directly above.
+unless File.exists?('/etc/opscode/pivotal.pem')
+  file "/etc/opscode/pivotal.pub" do
+    owner OmnibusHelper.new(node).ownership['owner']
+    group "root"
+    mode "0644"
+    content key.public_key.to_s
+  end
 end
 
+# If we are doing initial key generation (aka pivotal.pem hasn't yet
+# been created), create pivotal.pem.
 file "/etc/opscode/pivotal.pem" do
   owner OmnibusHelper.new(node).ownership['owner']
   group "root"